backend/generate_vocab_question/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.69.0
google-cloud-texttospeech==2.15.0
anthropic==0.42.0
flask==3.0.2
